+++ src/i16x16_.rs
@@ -72,6 +72,12 @@ impl Mul for i16x16 {
   }
 }
 
+integer_impl_div_rem!(
+  i16,
+  i16x16,
+  [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
+);
+
 impl Shl for i16x16 {
   type Output = Self;