+++ src/u32x16_.rs
@@ -121,6 +121,12 @@ impl Mul for u32x16 {
   }
 }
 
+integer_impl_div_rem!(
+  u32,
+  u32x16,
+  [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
+);
+
 impl BitAnd for u32x16 {
   type Output = Self;
   #[inline]