+++ src/u8x16_.rs
@@ -170,6 +170,12 @@ impl Mul for u8x16 {
   }
 }
 
+integer_impl_div_rem!(
+  u8,
+  u8x16,
+  [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
+);
+
 impl Shl for u8x16 {
   type Output = Self;