+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Added several functions and trait implementations that previously were only
   implemented for some types inconsistently.
+* Added integer `Div` and `Rem` implementations.
 * Added integer functions `clamp` and `saturating_mul`.
 * Added signed integer functions `is_positive` and `signum`.
 * Fixed slight bug in `f64xN` functions `round` and `round_int`.